+++ Locales/frFR.lua
@@ -1,0 +1,25 @@
+-- Dejunk: frFR (French) localization file.
+-- Translators: Arcueild
+
+local AddonName, AddonTable = ...
+local L = LibStub('AceLocale-3.0'):NewLocale(AddonName, 'frFR')
+if not L then return end
+
+L["AUTO_DESTROY_TEXT"] = "Destruction automatique"
+L["AUTO_DESTROY_TOOLTIP"] = "Détruis les objets indésirables régulièrement pendant que cette fenêtre est fermée."
+L["AUTO_REPAIR_TEXT"] = "Réparation automatique"
+L["AUTO_REPAIR_TOOLTIP"] = "Répare automatiquement les objets lors de l'ouverture d'une fenêtre de marchand."
+L["AUTO_SELL_TEXT"] = "Vente automatique"
+L["AUTO_SELL_TOOLTIP"] = "Vend automatiquement des objets indésirables lors de l'ouverture d'une fenêtre de marchand."
+L["BY_CATEGORY_TEXT"] = "Par catégorie"
+L["BY_QUALITY_TEXT"] = "Par qualité"
+L["CHARACTER_SPECIFIC_TEXT"] = "Paramètres spécifiques au personnage"
+L["CHARACTER_SPECIFIC_TOOLTIP"] = "Cliquez pour basculer entre les paramètres globaux et les paramètres spécifiques à ce personnage."
+L["EPIC_TEXT"] = "Epique"
+L["ITEM_WILL_BE_SOLD"] = "Cet objet sera vendu."
+L["ITEM_WILL_NOT_BE_SOLD"] = "Cet objet ne sera pas vendu."
+L["RARE_TEXT"] = "Rare"
+L["SELL_ALL_TOOLTIP"] = "Vendre tous les objets de cette qualité."
+L["SELL_TEXT"] = "Vendre"
+L["UNCOMMON_TEXT"] = "Inhabituel"
+L["VERBOSE_MODE_TOOLTIP"] = "Activer les messages supplémentaires Dejunk dans la fenêtre de discussion lors de la vente et la destruction d'objets."