+++ Modules/ListManager.lua
@@ -8,42 +8,36 @@ local DCL = Addon.Libs.DCL
 
 -- Upvalues
 local assert, pairs, next = assert, pairs, next
-local remove, sort, concat = table.remove, table.sort, table.concat
-local tonumber, tostring = tonumber, tostring
+local tremove, sort, concat = table.remove, table.sort, table.concat
+local format, select, tonumber, tostring = format, select, tonumber, tostring
+
+local GetItemInfo = GetItemInfo
 
 -- Modules
 local ListManager = Addon.ListManager
-
-local Core = Addon.Core
-local Colors = Addon.Colors
-local Tools = Addon.Tools
-local DB = Addon.DB
-
--- Variables
-ListManager.Initialized = false
-
-ListManager.Lists =
-{
+ListManager.Lists = {
   ["Inclusions"] = {},
   ["Exclusions"] = {},
   ["Destroyables"] = {}
-
   -- ["ListName"] = array of item tables
 }
-
 ListManager.ToAdd = {}
 ListManager.ToRemove = {}
-
--- Add list keys to ListManager, ToAdd, and ToRemove
+-- Add list keys to ToAdd and ToRemove
 for k in pairs(ListManager.Lists) do
-  ListManager[k] = k
   ListManager.ToAdd[k] = {}
   ListManager.ToRemove[k] = {}
 end
 
--- Parsing Frame
-local parseFrame = CreateFrame("Frame", AddonName.."ListManagerParseFrame")
-parseFrame.AttemptsToParse = {} -- Format: ["itemID"] = # of attempts
+local Core = Addon.Core
+local Colors = Addon.Colors
+local Dejunker = Addon.Dejunker
+local Destroyer = Addon.Destroyer
+local Tools = Addon.Tools
+local DB = Addon.DB
+
+-- Consts
+local MAX_NUMBER = 2147483647 -- 32-bit signed
 
 -- ============================================================================
 -- General Functions
@@ -88,13 +82,13 @@ end
 -- Gets a list from saved variables.
 -- @param listName - the name of the list to get SVs
 function ListManager:GetListSV(listName)
-  assert(self[listName])
+  assert(self.Lists[listName])
   return DB.Profile[listName]
 end
 
 -- Returns list data for use in ListManager functions.
 function ListManager:GetListData(listName)
-  assert(self[listName])
+  assert(self.Lists[listName])
 
   local list = self.Lists[listName]
   local toAdd = self.ToAdd[listName]
@@ -103,10 +97,10 @@ function ListManager:GetListData(listName)
   local coloredName = Tools:GetColoredListName(listName)
 
   local otherListName = nil
-  if (listName == self.Inclusions) then
-    otherListName = self.Exclusions
-  elseif (listName == self.Exclusions) then
-    otherListName = self.Inclusions
+  if (listName == "Inclusions") then
+    otherListName = "Exclusions"
+  elseif (listName == "Exclusions") then
+    otherListName = "Inclusions"
   end
 
   return list, toAdd, toRemove, sv, coloredName, otherListName
@@ -120,7 +114,7 @@ end
 -- @param listName - the name of the list to add to
 -- @param itemID - the item id of the item to add
 function ListManager:AddToList(listName, itemID)
-  assert(self[listName] ~= nil)
+  assert(self.Lists[listName] ~= nil)
   itemID = tostring(itemID)
 
   -- Initialize data
@@ -149,7 +143,7 @@ end
 -- @param itemID - the item id of the item to remove
 -- @param notify - if true, prints a message if the item is not on the list
 function ListManager:RemoveFromList(listName, itemID, notify)
-  assert(self[listName] ~= nil)
+  assert(self.Lists[listName] ~= nil)
   itemID = tostring(itemID)
 
   if self:IsOnList(listName, itemID) then
@@ -168,28 +162,16 @@ end
 -- @param itemID - the item id of the item to search for
 -- @return - boolean
 function ListManager:IsOnList(listName, itemID)
-  assert(self[listName] ~= nil)
+  assert(self.Lists[listName] ~= nil)
   itemID = tostring(itemID)
 
   return self:GetListSV(listName)[itemID]
 end
 
--- Sorts the specified item list.
--- @param list - a table of items to sort
-function ListManager:SortList(list)
-  -- Sort by name when same quality, otherwise sort by quality.
-  sort(list, function(a, b)
-    if ((a.Quality - b.Quality) == 0) then
-      return (a.Name < b.Name) end
-
-    return (a.Quality < b.Quality)
-  end)
-end
-
 -- Removes all entries from the specified list.
 -- @param listName - the name of the list to destroy
 function ListManager:DestroyList(listName)
-  assert(self[listName] ~= nil)
+  assert(self.Lists[listName] ~= nil)
 
   -- Initialize data
   local list, _, _, sv, coloredName = self:GetListData(listName)
@@ -209,11 +191,11 @@ end
 -- @param listName - the name of the list to import into
 -- @param string - the import string with format: "itemID;itemID;itemID;"
 function ListManager:ImportToList(listName, string)
-  assert(self[listName] ~= nil)
+  assert(self.Lists[listName] ~= nil)
 
   for itemID in string:gmatch('([^;]+)') do
     itemID = tonumber(itemID)
-    if (itemID and (itemID > 0)) then
+    if (itemID and (itemID > 0) and (itemID <= MAX_NUMBER)) then
       self:AddToList(listName, itemID)
     end
   end
@@ -223,7 +205,7 @@ end
 -- @param listName - the name of the list to export
 -- @return - a string with format: "itemID;itemID;itemID;"
 function ListManager:ExportFromList(listName)
-  assert(self[listName] ~= nil)
+  assert(self.Lists[listName] ~= nil)
 
   -- Initialize data
   local sv = self:GetListSV(listName)
@@ -240,29 +222,27 @@ end
 -- Parsing Functions
 -- ============================================================================
 
--- Set OnUpdate script
-function parseFrame:OnUpdate(elapsed)
-  if Addon.Dejunker:IsDejunking() then return end
+-- OnUpdate(), called in Core:OnUpdate()
+function ListManager:OnUpdate(elapsed)
+  if Dejunker:IsDejunking() or Destroyer:IsDestroying() then return end
 
   -- Removals
-  for listName, list in pairs(ListManager.ToRemove) do
-    if next(list) then ListManager:CleanList(listName) end
+  for listName, list in pairs(self.ToRemove) do
+    if next(list) then self:CleanList(listName) end
   end
 
   -- Additions
-  for listName, list in pairs(ListManager.ToAdd) do
-    if next(list) then ListManager:ParseList(listName) end
+  for listName, list in pairs(self.ToAdd) do
+    if next(list) then self:ParseList(listName) end
   end
 end
-
-parseFrame:SetScript("OnUpdate", parseFrame.OnUpdate)
 
 do -- CleanList()
   -- Removes an item from the specified list by ID.
   local function removeItem(list, itemID, sv, coloredName)
     for k, v in pairs(list) do
       if (v.ItemID == itemID) then
-        remove(list, k) -- remove has to be used here for the table to update as expected
+        tremove(list, k) -- tremove has to be used here for the table to update as expected
         sv[itemID] = nil
         Core:Print(format(L.REMOVED_ITEM_FROM_LIST, v.ItemLink, coloredName))
         return
@@ -273,7 +253,7 @@ do -- CleanList()
   -- Cleans the specified list by removing queued entries.
   -- @param listName - the name of the list to clean
   function ListManager:CleanList(listName)
-    assert(self[listName])
+    assert(self.Lists[listName])
 
     -- Initialize data
     local list, _, toRemove, sv, coloredName = self:GetListData(listName)
@@ -290,6 +270,9 @@ end
 
 do -- ParseList()
   local MAX_PARSE_ATTEMPTS = 100
+  local parseAttempts = {
+    -- [itemID] = count
+  }
   
   -- Creates and returns an item by item id.
   -- @param itemID - the item id of the item to create
@@ -311,7 +294,7 @@ do -- ParseList()
 
   -- Returns true if the item can be sold, and the target list is Inclusions or Exclusions.
   local function canBeSold(listName, item, sv, toAdd)
-    if not (listName == ListManager.Inclusions or listName == ListManager.Exclusions) then  return false end
+    if not (listName == "Inclusions" or listName == "Exclusions") then  return false end
     if Tools:ItemCanBeSold(item) then return true end
     -- Item cannot be sold
     sv[item.ItemID] = nil -- Remove from sv
@@ -322,7 +305,7 @@ do -- ParseList()
 
   -- Returns true if the item can be destroyed, and the target list is Destroyables.
   local function canBeDestroyed(listName, item, sv, toAdd)
-    if not (listName == ListManager.Destroyables) then return false end
+    if not (listName == "Destroyables") then return false end
     if Tools:ItemCanBeDestroyed(item) then return true end
     -- Item cannot be destroyed
     sv[item.ItemID] = nil -- Remove from sv
@@ -343,7 +326,7 @@ do -- ParseList()
   -- Parses queued itemIDs and adds them to the specified list.
   -- @param listName - the name of the list to parse
   function ListManager:ParseList(listName)
-    assert(self[listName])
+    assert(self.Lists[listName])
 
     -- Initialize data
     local list, toAdd, _, sv, coloredName = self:GetListData(listName)
@@ -365,17 +348,17 @@ do -- ParseList()
         end
 
         list[#list+1] = item
+        parseAttempts[itemID] = nil
         toAdd[itemID] = nil
       elseif not sv[itemID] then -- if the item couldn't be parsed, and it is not in the saved variables (avoids erasing sv data by accident)
-        local attempts = ((parseFrame.AttemptsToParse[itemID] or 0) + 1)
-
+        local attempts = (parseAttempts[itemID] or 0) + 1
         if (attempts >= MAX_PARSE_ATTEMPTS) then
-          attempts = 0
+          parseAttempts[itemID] = nil
           toAdd[itemID] = nil
           Core:Print(format(L.FAILED_TO_PARSE_ITEM_ID, DCL:ColorString(itemID, DCL.CSS.Grey)))
+        else
+          parseAttempts[itemID] = attempts
         end
-
-        parseFrame.AttemptsToParse[itemID] = attempts
       end
     end
 
@@ -384,9 +367,7 @@ do -- ParseList()
       sort(list, item_sort)
 
       -- Start auto destroy if the Destroyables list was updated
-      if (listName == ListManager.Destroyables) then
-        Addon.Destroyer:StartAutoDestroy()
-      end
+      if (listName == "Destroyables") then Destroyer:StartAutoDestroy() end
     end
   end
 end