+++ CHANGELOG.md
@@ -1,2 +1,5 @@
-### 8.0.5
-* Updated libraries
+### 8.0.6
+* Latency is now taken into account when selling and destroying items.
+* Ignore Binds When Equipped no longer ignores poor quality items.
+* Localization updates
+* Code improvements